// frontend/src/api/generated/models/timeSlotResponse.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * 途排智策 API
 * 途排智策前后端分离 MVP API
 * OpenAPI spec version: 0.1.0
 */

export interface TimeSlotResponse {
  campus_id: string;
  business_id: string;
  weekday: string;
  start_time: string;
  end_time: string;
  kind?: string;
  sequence?: number;
  is_open?: boolean;
  id: string;
}